move APIFnSign to action_sign.go
